+++ lib/config/router/router.dart
@@ -1,4 +1,4 @@
-import 'package:base_project/presentation/widgets/widget_tree.dart';
+import 'package:base_project/presentation/widget_tree.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import '../../screens/index.dart';
@@ -15,13 +15,13 @@ final GoRouter router = GoRouter(
     GoRoute(
       path: '/login',
       builder: (BuildContext context, GoRouterState state) {
-        return const LoginPage();
+        return const LoginScreen();
       },
     ),
     GoRoute(
-      path: '/test',
+      path: '/calendar',
       builder: (BuildContext context, GoRouterState state) {
-        return Test();
+        return CalendarScreen();
       },
     ),
   ],