+++ lib/screens/login/login_screen.dart
@@ -1,19 +1,19 @@
 import 'package:flutter/material.dart';
-import 'package:base_project/presentation/widgets/input_text_box.dart';
-import 'package:go_router/go_router.dart';
+import 'package:base_project/presentation/widgets/inputs/input_text_box.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import '../config/auth/auth.dart';
+import '../../config/auth/auth.dart';
 
-class LoginPage extends StatefulWidget {
-  const LoginPage({super.key});
+class LoginScreen extends StatefulWidget {
+  const LoginScreen({super.key});
 
   @override
-  State<LoginPage> createState() => _LoginPageState();
+  State<LoginScreen> createState() => _LoginScreenState();
 }
 
-class _LoginPageState extends State<LoginPage> {
+class _LoginScreenState extends State<LoginScreen> {
   String? errorMessage = '';
   bool isLogin = true;
+  bool isPasswordHidden = true;
 
   final emailController = TextEditingController();
   final passwordController = TextEditingController();
@@ -64,16 +64,17 @@ class _LoginPageState extends State<LoginPage> {
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               Text('Food Structured',
-                  style: TextStyle(
-                      fontFamily: 'Roboto',
-                      fontSize: 40.0,
-                      color: colors.primary)),
+                style: TextStyle(
+                  fontFamily: 'Roboto',
+                  fontSize: 40.0,
+                  color: colors.primary
+                ),
+              ),
               const SizedBox(height: 15.0),
               const CircleAvatar(
                 radius: 100.0,
                 backgroundColor: Colors.white,
-                backgroundImage:
-                    AssetImage('resources/images/FoodStructuredLogo.png'),
+                backgroundImage: AssetImage('resources/images/FoodStructuredLogo.png'),
               ),
               const SizedBox(height: 15.0),
               Text(
@@ -90,28 +91,25 @@ class _LoginPageState extends State<LoginPage> {
               InputTextBox(
                 controller: passwordController,
                 label: 'Contraseña',
-                icon: Icons.lock,
-                hideText: true,
+                suffixIcon: IconButton(
+                  icon: Icon(isPasswordHidden ? Icons.visibility : Icons.visibility_off), 
+                  onPressed: () => setState(() => isPasswordHidden = !isPasswordHidden),
+                ),
+                hideText: isPasswordHidden,
               ),
               TextButton(
                 style: ButtonStyle(
-                    foregroundColor:
-                        MaterialStateProperty.all<Color>(Colors.blue)),
-                onPressed: () {
-                  setState(() {
-                    isLogin = !isLogin;
-                  });
-                },
-                child: Text(isLogin
-                    ? 'Cambiar a registro'
-                    : 'Cambiar a inicio de sesión'),
+                  foregroundColor: MaterialStateProperty.all<Color>(colors.primary),
+                ),
+                onPressed: () => setState(() => isLogin = !isLogin),
+                child: Text(isLogin ? 'Cambiar a registro' : 'Cambiar a inicio de sesión'),
               ),
-              const SizedBox(height: 15.0),
               SizedBox(
                 child: TextButton(
                   style: ButtonStyle(
-                      foregroundColor:
-                          MaterialStateProperty.all<Color>(Colors.blue)),
+                    foregroundColor: MaterialStateProperty.all<Color>(Colors.white),
+                    backgroundColor: MaterialStateProperty.all<Color>(colors.primary),
+                  ),
                   onPressed: () {
                     if (isLogin) {
                       signInWithEmailAndPassword();
@@ -122,8 +120,7 @@ class _LoginPageState extends State<LoginPage> {
                   child: Text(isLogin ? 'Iniciar Sesión' : 'Registrarse'),
                 ),
               ),
-              Text(errorMessage ?? '',
-                  style: const TextStyle(color: Colors.red)),
+              Text(errorMessage ?? '', style: const TextStyle(color: Colors.red)),
             ],
           )
         ],