+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This file is used to explain in detail changes made to the food structured app.
 <!-- TOC -->
   [[TOC]]
 
+## V 1.X.X
+Date: Nov XX, 2023
+* [NEW]
+  * Blank calendar screen and drawer menu
+
 ## V 1.0.0
 Date: Nov 05, 2023
 * [NEW]